modified start scripts and updated readme
