+++ start.bat
@@ -1,2 +1,3 @@
-pip install -r rquirements.txt
+pytest -p no:warnings tests/test_environment.py
+pause
 python src\main.py
